src/loadgenerator/requirements.txt
backoff==2.2.1
brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
configargparse==1.5.3
deprecated==1.2.13
flask==2.2.2
flask-basicauth==0.2.0
flask-cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
googleapis-common-protos==1.57.0
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
jinja2==3.1.2
locust==2.14.0
markupsafe==2.1.1
msgpack==1.0.4
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-urllib3==0.36b0
opentelemetry-propagator-b3==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
protobuf==4.21.12
psutil==5.9.4
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.13
werkzeug==2.2.2
wrapt==1.14.1
zope-event==4.6
zope-interface==5.5.2
